EXAM2/project/rooms/room.py
from project.appliances.appliance import Appliance


class Room:
    def __init__(self, name: str, budget: float, members_count: int):
        self.family_name = name
        self.budget = budget
        self.members_count = members_count
        self.children = []
        self.expenses = 0

    @property
    def expenses(self):
        return self._expenses

    @expenses.setter
    def expenses(self, value):
        if value < 0:
            raise ValueError("Expenses cannot be negative")
        self._expenses = value

    def calculate_expenses(self, *args):
        total = 0
        for arg in args:
            for a in arg:
                if isinstance(a, Appliance):
                    total += a.get_monthly_expense()
                else:
                    total += a.cost * 30

        self.expenses = total
